update README for latest build script
I've learned a lot the past few days, and my initial idea for how this
would work has been changed some. It won't be quite as idiot-proof as
I'd hoped, but it's not too bad. This updates the readme to reflect my
current understanding of how things will work here.