possible final build.sh
I think I've made it through everything, including the optional
widescreen support, except for the better cutscenes. I had downloaded
the 60fps of those, and the instructions say to use 30, so I'm waiting
on that to download again. It only replaces movie files, though, so I
can start a fresh build from scratch and see how it goes.