prep to commit mod files with large mods split
I had previously attempted to push up the mod archives, but some are too
large. I've completed the install script, and now it's time to figure
out the mod storage. I've successfully split up all archives larger than
25M, added them to the .gitignore, and recorded sha1sums of them in two
forms: all in one `hashes` file and individually in files in
`checksums`. I've also written an `unsplit.sh` script that takes on of
the folders in `mod-archives-splits` and reassembles it with checksum
verification. At this point, I should just need to commit all the mods
and chunks, and then write up a final script (or chunk in `build.sh`)
that will unsplit ensure the mods are unsplit before trying to install
them.